feat: commit

+++ fofa/count.go
@@ -3,7 +3,7 @@ package fofa
 import "context"
 
 // Count 执行搜索查询并返回结果数量
-func (c *Client) Count(ctx context.Context, query string) (int64, error) {
+func (c *Client) Count(ctx context.Context, query string) (int, error) {
 	req := &SearchRequest{
 		Ctx:   ctx,
 		Query: query,
@@ -15,5 +15,5 @@ func (c *Client) Count(ctx context.Context, query string) (int64, error) {
 	if err != nil {
 		return 0, err
 	}
-	return int64(resp.Size), nil
+	return resp.Size, nil
 }
